+++ Editor/d4rkAvatarOptimizer.cs
@@ -770,7 +770,7 @@ public class d4rkAvatarOptimizer : MonoBehaviour
             for (int i = 0; fxLayer != null && i < fxLayer.layers.Length; i++) {
                 if (fxLayer.layers[i] == null || fxLayer.layers[i].stateMachine == null)
                     continue;
-                if (uselessLayers.Contains(i) || IsMergeableFXLayer(i))
+                if (OptimizeFXLayer && (uselessLayers.Contains(i) || IsMergeableFXLayer(i)))
                     continue;
                 goOffPaths.Clear();
                 goOnPaths.Clear();